+++ packages/web-core/src/entities/capability.entity.ts
@@ -14,7 +14,9 @@ export class Capability {
       logsDebug = false,
       narratedActions = false,
       multiModalActionPlanning: multiModalActionPlanning = false,
+      perceivedLatencyReport: perceivedLatencyReport = true,
       phonemes: phonemeInfo = false,
+      pingPongReport: pingPongReport = true,
       silence: silenceEvents = false,
     } = capabilities;
 
@@ -30,7 +32,9 @@ export class Capability {
       multiAgent: true,
       multiModalActionPlanning,
       narratedActions,
+      perceivedLatencyReport,
       phonemeInfo,
+      pingPongReport,
       silenceEvents,
     };
   }